+++ src/main/java/racingcar/domain/car/Car.java
@@ -1,4 +1,6 @@
-package racingcar.domain;
+package racingcar.domain.car;
+
+import racingcar.domain.movestrategy.MoveStrategy;
 
 public class Car {
     private final Name name;
