+++ src/main/java/racingcar/domain/movestrategy/MoveStrategy.java
@@ -1,4 +1,6 @@
-package racingcar.domain;
+package racingcar.domain.movestrategy;
+
+import racingcar.domain.car.Car;
 
 public interface MoveStrategy {
     boolean isAvailableMove(final Car car);
