+++ src/main/java/racingcar/service/dto/CarDto.java
@@ -1,6 +1,6 @@
 package racingcar.service.dto;
 
-import racingcar.domain.Car;
+import racingcar.domain.car.Car;
 
 import java.util.Objects;
