+++ src/test/java/racingcar/domain/car/NameTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.car;
 
 import org.junit.jupiter.api.Test;
