+++ src/test/java/racingcar/domain/race/RaceResultTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.race;
 
 import org.junit.jupiter.api.Test;
 import racingcar.service.dto.CarDto;
