+++ src/test/java/racingcar/domain/race/RaceStatusTest.java
@@ -1,4 +1,4 @@
-package racingcar.domain;
+package racingcar.domain.race;
 
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
